m/pip-requirements.txt
ruamel.yaml==0.13.11
jinja2==2.9
fire==0.1.0
pip>=0.9
